JB-20: fixes saving date firebase part 1

+++ src/services/deleteItemFromFirebase.tsx
@@ -1,9 +1,27 @@
 import { db } from '../firebase';
 import { doc, deleteDoc } from 'firebase/firestore';
+import { getAuth } from 'firebase/auth';
 
 export const deleteItemFromFirebase = async (category, itemId) => {
   try {
-    const itemRef = doc(db, 'categories', category, 'products', itemId);
+    const auth = getAuth();
+    const currentUser = auth.currentUser;
+
+    if (!currentUser) {
+      throw new Error('User not authenticated');
+    }
+
+    const userId = currentUser.uid;
+
+    const itemRef = doc(
+      db,
+      'users',
+      userId,
+      'categories',
+      category,
+      'products',
+      itemId,
+    );
 
     await deleteDoc(itemRef);
